requirements.txt
fastapi 
uvicorn 
psycopg2-binary 
sqlalchemy 
python-jose 
passlib
pydantic[email]
bcrypt==4.0.1